requirements.txt
astroid==2.3.2
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.1
pylint==2.4.3
sentencepiece==0.1.85
torch==1.4.0
translate-model==0.1.0
typed-ast==1.4.0
Werkzeug==1.0.1
wrapt==1.11.2
